Added Practitioner Rating Calculation
Added the calculation of practitioner ratings to the practitioner homepage and search.

+++ MediLink/Entities/Practitioner.cs
@@ -57,7 +57,7 @@ namespace MediLink.Entities
 
 		public PractitionerType PractitionerType { get; set; } = null!;
 
-        public int? rating { get; set; }
+        public double? rating { get; set; }
 
     }
 }
